Use a primary namespace for OpenFaaS CE
See notes on intended use for OpenFaaS CE in the docs:

https://docs.openfaas.com/openfaas-pro/introduction/
