+++ .github/workflows/publish.yaml
@@ -7,11 +7,7 @@ on:
 
 jobs:
   publish:
-    strategy:
-      matrix:
-        go-version: [1.18.x]
-        os: [ubuntu-latest]
-    runs-on: ${{ matrix.os }}
+    runs-on: ubuntu-latest
     permissions:
       actions: read
       checks: write
@@ -25,7 +21,7 @@ jobs:
       - name: Install Go
         uses: actions/setup-go@v3.1.0
         with:
-          go-version: ${{ matrix.go-version }}
+          go-version: 1.18.x
       - name: Set up QEMU
         uses: docker/setup-qemu-action@v2
       - name: Set up Docker Buildx
