+++ pkg/config/read_config.go
@@ -47,9 +47,8 @@ func (ReadConfig) Read(hasEnv ftypes.HasEnv) (BootstrapConfig, error) {
 		return cfg, fmt.Errorf("invalid image_pull_policy configured: %s", imagePullPolicy)
 	}
 
-	cfg.DefaultFunctionNamespace = ftypes.ParseString(hasEnv.Getenv("function_namespace"), "default")
+	cfg.DefaultFunctionNamespace = ftypes.ParseString(hasEnv.Getenv("function_namespace"), "openfaas-fn")
 	cfg.ProfilesNamespace = ftypes.ParseString(hasEnv.Getenv("profiles_namespace"), cfg.DefaultFunctionNamespace)
-	cfg.ClusterRole = ftypes.ParseBoolValue(hasEnv.Getenv("cluster_role"), false)
 
 	cfg.HTTPProbe = httpProbe
 	cfg.SetNonRootUser = setNonRootUser
@@ -117,9 +116,6 @@ type BootstrapConfig struct {
 
 	// FaaSConfig contains the configuration for the FaaSProvider
 	FaaSConfig ftypes.FaaSConfig
-
-	// ClusterRole determines whether the operator should have cluster wide access
-	ClusterRole bool
 }
 
 // Fprint pretty-prints the config with the stdlib logger. One line per config value.
@@ -143,6 +139,5 @@ func (c BootstrapConfig) Fprint(verbose bool) {
 		log.Printf("LivenessProbeInitialDelaySeconds: %d\n", c.LivenessProbeInitialDelaySeconds)
 		log.Printf("LivenessProbeTimeoutSeconds: %d\n", c.LivenessProbeTimeoutSeconds)
 		log.Printf("LivenessProbePeriodSeconds: %d\n", c.LivenessProbePeriodSeconds)
-		log.Printf("ClusterRole: %v\n", c.ClusterRole)
 	}
 }
