+++ pkg/handlers/namespaces.go
@@ -19,20 +19,14 @@ import (
 )
 
 // MakeNamespacesLister builds a list of namespaces with an "openfaas" tag, or the default name
-func MakeNamespacesLister(defaultNamespace string, clusterRole bool, clientset kubernetes.Interface) http.HandlerFunc {
+func MakeNamespacesLister(defaultNamespace string, clientset kubernetes.Interface) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		// log.Println("List namespaces")
 
 		if r.Body != nil {
 			defer r.Body.Close()
 		}
 
-		namespaces := []string{}
-		if clusterRole {
-			namespaces = ListNamespaces(defaultNamespace, clientset)
-		} else {
-			namespaces = append(namespaces, defaultNamespace)
-		}
+		namespaces := []string{defaultNamespace}
 
 		out, err := json.Marshal(namespaces)
 		if err != nil {
@@ -66,6 +60,10 @@ func NewNamespaceResolver(defaultNamespace string, kube kubernetes.Interface) Na
 				req.Namespace = namespace
 			}
 
+			if req.Namespace != defaultNamespace {
+				return "", fmt.Errorf("namespace %s is not allowed", req.Namespace)
+			}
+
 		case http.MethodPost, http.MethodPut, http.MethodDelete:
 			body, _ := ioutil.ReadAll(r.Body)
 			err := json.Unmarshal(body, &req)
@@ -74,17 +72,19 @@ func NewNamespaceResolver(defaultNamespace string, kube kubernetes.Interface) Na
 				return "", fmt.Errorf("unable to unmarshal json request")
 			}
 
+			if req.Namespace != defaultNamespace {
+				return "", fmt.Errorf("namespace %s is not allowed", req.Namespace)
+			}
+
 			// Reconstruct Body
 			r.Body = ioutil.NopCloser(bytes.NewBuffer(body))
 		}
 
-		allowedNamespaces := ListNamespaces(defaultNamespace, kube)
-		ok := findNamespace(req.Namespace, allowedNamespaces)
-		if !ok {
-			return req.Namespace, fmt.Errorf("unable to manage secrets within the %s namespace", req.Namespace)
+		if req.Namespace != defaultNamespace {
+			return "", fmt.Errorf("unable to manage secrets within the %s namespace", req.Namespace)
 		}
 
-		return req.Namespace, nil
+		return defaultNamespace, nil
 	}
 }
